requirements.txt
dateutils==0.6.12
python-dateutil~=2.8.2
requests~=2.31.0
pillow~=10.1.0